+++ popkorn-compiler/src/main/kotlin/cc/popkorn/compiler/PopKornException.kt
@@ -4,6 +4,6 @@ package cc.popkorn.compiler
  * Exceptions used when compiling PopKorn
  *
  * @author Pau Corbella
- * @since 1.0
+ * @since 1.0.0
  */
 internal class PopKornException(msg:String) : RuntimeException(msg)