+++ popkorn/src/main/kotlin/cc/popkorn/PopKornController.kt
@@ -6,7 +6,7 @@ import kotlin.reflect.KClass
  * Interface with the available methods to manage injections
  *
  * @author Pau Corbella
- * @since 1.0
+ * @since 1.0.0
  */
 interface PopKornController {
 
@@ -15,5 +15,7 @@ interface PopKornController {
      fun <T:Any> removeInjectable(type:KClass<T>, environment:String?=null)
 
      fun reset()
+
+     fun purge()
 
 }