+++ popkorn/src/main/kotlin/cc/popkorn/core/Injector.kt
@@ -17,12 +17,12 @@ import org.apache.commons.io.IOUtils
 
 
 /**
- * Main class to perform the injections
+ * Main class to perform injections
  *
  * @author Pau Corbella
- * @since 1.0
+ * @since 1.0.0
  */
-internal class Injector(private val debug:Boolean=false) : PopKornController {
+class Injector(private val debug:Boolean=false) : PopKornController {
     private val resolverPool = ResolverPool()
     private val providerPool = ProviderPool()
 
@@ -86,12 +86,22 @@ internal class Injector(private val debug:Boolean=false) : PopKornController {
 
 
     /**
+     * Frees memory that is not needed anymore
+     *
+     * */
+    override fun purge(){
+        instances.mapNotNull { it.value as? VolatileInstances }
+            .forEach { it.purge() }
+    }
+
+
+    /**
      * Retrieves an object of type clazz (it will be created or provided depending on its Scope)
      *
      * @param clazz Class or Interface that you want to retrieve
      * @param environment The environment in which you would like to retrieve the object
      */
-    fun <T:Any> inject(clazz: KClass<T>, environment:String?) : T {
+    fun <T:Any> inject(clazz: KClass<T>, environment:String?=null) : T {
         return if (clazz.isInterface()){
             val impl = clazz.getImplementation(environment)
             provide(impl, environment)
@@ -115,9 +125,9 @@ internal class Injector(private val debug:Boolean=false) : PopKornController {
     private fun <T: Any> KClass<T>.getInstances() : Instances<T>{
         val provider = providerPool.create(this)
         return when(provider.scope()){
-            Scope.BY_APP -> PersistentInstances(provider)
-            Scope.BY_USE -> VolatileInstances(provider)
-            Scope.BY_NEW -> NewInstances(provider)
+            Scope.BY_APP -> PersistentInstances(this@Injector, provider)
+            Scope.BY_USE -> VolatileInstances(this@Injector, provider)
+            Scope.BY_NEW -> NewInstances(this@Injector, provider)
         }
     }