+++ popkorn/src/main/kotlin/cc/popkorn/core/Provider.kt
@@ -7,11 +7,11 @@ import cc.popkorn.Scope
  * T cannot be an interface, only classes
  *
  * @author Pau Corbella
- * @since 1.0
+ * @since 1.0.0
  */
 interface Provider<T:Any> {
 
-     fun create(environment:String?) : T
+     fun create(injector:Injector, environment:String?) : T
 
      fun scope() : Scope