+++ popkorn/src/main/kotlin/cc/popkorn/instances/NewInstances.kt
@@ -1,5 +1,6 @@
 package cc.popkorn.instances
 
+import cc.popkorn.core.Injector
 import cc.popkorn.core.Provider
 
 
@@ -8,12 +9,12 @@ import cc.popkorn.core.Provider
  * Doesn't matter the number of times get() is called, it will always return a new instance
  *
  * @author Pau Corbella
- * @since 1.0
+ * @since 1.0.0
  */
-internal class NewInstances<T:Any>(private val provider: Provider<T>): Instances<T> {
+internal class NewInstances<T:Any>(private val injector: Injector, private val provider: Provider<T>): Instances<T> {
 
     override fun get(environment:String?) : T{
-        return provider.create(environment)
+        return provider.create(injector, environment)
     }
 
     override fun size() = 0