+++ popkorn/src/main/kotlin/cc/popkorn/instances/VolatileInstances.kt
@@ -1,5 +1,6 @@
 package cc.popkorn.instances
 
+import cc.popkorn.core.Injector
 import cc.popkorn.core.Provider
 import java.lang.ref.WeakReference
 
@@ -10,15 +11,20 @@ import java.lang.ref.WeakReference
  * this instance is being used by others. If no other object is using it then it will create a new one
  *
  * @author Pau Corbella
- * @since 1.0
+ * @since 1.0.0
  */
-internal class VolatileInstances<T:Any>(private val provider: Provider<T>): Instances<T> {
+internal class VolatileInstances<T:Any>(private val injector: Injector, private val provider: Provider<T>): Instances<T> {
     private val instances = HashMap<String?, WeakReference<T>>()
-    
+
+    @Synchronized
     override fun get(environment:String?) : T{
-        return instances[environment]?.get() ?: provider.create(environment).also { instances[environment] = WeakReference(it) }
+        return instances[environment]?.get() ?: provider.create(injector, environment).also { instances[environment] = WeakReference(it) }
     }
 
+    @Synchronized
     override fun size() = instances.size
+
+    @Synchronized
+    fun purge() = instances.filter { it.value.get()==null }.forEach { instances.remove(it.key) }
 
 }